requirements.txt
torch
torchvision
torchaudio
groundingdino-py
opencv-python
pillow
matplotlib
